Transaction / Order(s) / OrderStatus / Fill: adjust toString
CombinationServiceImpl / PositionServiceImpl / TransactionServiceImpl: improved logging

TradeCallback: add owningStrategyName to alias

EserManager:
add processing ThreadLocal
add internalSendEvent
remove routeEvent (replace with sendEvent calls)
sendExternalEvent -> externalSendEvent
addTradeCallback: add owningStrategyName to alias

module-prepared.epl (ON_TRADE_COMPLETED): add owningStrategyName to alias

module-portfolio.epl (CLOSE_POSITION): add "group by" and "output"

git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@5161 19eef36c-86f0-de11-afab-0004ac968479

+++ core/src/main/java/com/algoTrader/service/CombinationServiceImpl.java
@@ -177,19 +177,22 @@ public class CombinationServiceImpl extends CombinationServiceBase {
             return;
         }
 
-        logger.info("close all positions of combination " + combination + " and delete combination");
-
         // reduce all associated positions by the specified amount
         // Note: positions are not closed, because other combinations might relate to them as well
         for (Component component : combination.getComponents()) {
 
             Position position = getPositionDao().findBySecurityAndStrategy(component.getSecurity().getId(), strategyName);
+
+            logger.info("close position " + position.getId() + " of combination " + combination);
+
             getPositionService().reducePosition(position.getId(), component.getQuantity());
         }
 
-        // close all non-tradeable positions based on the combination
+        // close non-tradeable position on the combination
         Position position = getPositionDao().findBySecurityAndStrategy(combinationId, strategyName);
-        getPositionService().deleteNonTradeablePosition(position.getId(), true);
+        if (position != null) {
+            getPositionService().deleteNonTradeablePosition(position.getId(), true);
+        }
 
         // delete the combination
         deleteCombination(combination.getId());
